requirements.txt
click==8.1.7
ffmpeg-python==0.2.0
future==1.0.0
PyYAML==6.0.1
